+++ src/components/FrameworkContentPage/index.tsx
@@ -1,17 +1,24 @@
 
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useAsyncEffect } from 'ahooks';
 import { getContentBySlug } from '../../utils/framework/framework';
+import { SiderDataType } from '../../store/framework';
+import useWebPlaygroundStore from '../../store';
 
 
 export interface LayoutProps {
   slug: string[];
-  frontmatter: Record<string, any>;
+  siderData: SiderDataType[];
 }
 
 const FrameworkLayout = (props: LayoutProps) => {
-  const { slug, frontmatter } = props
+  const { slug, siderData } = props
   const [element, setElement] = useState<React.ReactNode>()
+  const updateSiderData = useWebPlaygroundStore(state => state.updateSiderData)
+
+  useEffect(() => {
+    updateSiderData(siderData);
+  }, [])
 
   useAsyncEffect(async () => {
     const { default: Component } = await getContentBySlug(slug)
