+++ src/components/Layout/framework/index.tsx
@@ -19,8 +19,9 @@ const FrameworkLayout = (props: any) => {
       <Layout>
         <Sider width={200} className={`site-layout-background navbar`}>
           <Menu
+            selectable={false}
             items={siderData.map(item => ({
-              label: <Link href={item.linkTo}>{item.title}</Link>,
+              label: <Link href={item.linkTo} >{item.title}</Link>,
               key: item.id
             }))}
           />
