+++ src/pages/framework/[...slug].tsx
@@ -2,6 +2,7 @@ import glob from 'fast-glob'
 import FrameworkContentPage from '../../components/FrameworkContentPage'
 import { GetStaticPaths, GetStaticProps } from 'next'
 import { getFrontmatterBySlug } from '../../utils/framework/framework'
+import { getFrameworkSiderData } from '../../utils/framework/sidebarData'
 
 export default FrameworkContentPage
 
@@ -18,7 +19,8 @@ export const getStaticPaths: GetStaticPaths<{ slug: string[] }> = async () => {
 
 export const getStaticProps: GetStaticProps<any, { slug: string[] }> = async (context) => {
   const { slug } = context.params!
-  const { data: frontmatter, content } = await getFrontmatterBySlug(slug)
+  const { data: frontmatter } = await getFrontmatterBySlug(slug)
+  const siderData = await getFrameworkSiderData();
 
-  return { props: { slug, frontmatter } }
+  return { props: { slug, frontmatter, siderData } }
 }
