+++ src/pages/framework/index.tsx
@@ -3,33 +3,14 @@ import React, { useEffect } from 'react'
 import { GetStaticProps } from 'next';
 import { getFrameworkSiderData } from '../../utils/framework/sidebarData';
 import useWebPlaygroundStore from '../../store';
-import matter from 'gray-matter';
 import { SiderDataType } from '../../store/framework';
 
-
-function formatSiderData(siderData: { filename: string; fileContent: string }[]) {
-  const formattedSiderData = siderData.map((item) => {
-    const { data: frontmatter } = matter(item.fileContent);
-    if (frontmatter.title) {
-      return {
-        id: Math.random(),
-        title: frontmatter.title,
-        linkTo: item.filename
-      }
-    }
-  })
-  return formattedSiderData;
-}
-const Framework = (prop: { siderData: { filename: string; fileContent: string }[] }) => {
+const Framework = (prop: { siderData: SiderDataType[] }) => {
   const { siderData } = prop
   const updateSiderData = useWebPlaygroundStore(state => state.updateSiderData)
 
   useEffect(() => {
-    (function () {
-      const formattedSiderData: SiderDataType[] = formatSiderData(siderData).filter(i => i !== undefined) as SiderDataType[];
-      console.log('siderData', siderData, formattedSiderData);
-      updateSiderData(formattedSiderData);
-    })();
+    updateSiderData(siderData);
   }, [])
 
   return <>
